Split PlantDefinition out
