+++ pvzclass/Classes.hpp
@@ -1,4 +1,6 @@
 ﻿#pragma once
+#include "Classes/Definition.hpp"
+
 #include "Classes/ChallengeScreen.hpp"
 #include "Classes/SeedChooserScreen.hpp"
 #include "Classes/TodParticleSystem.hpp"
